Update package versions in EasyDbMigrator and test projects
Updated the version of the `EasyDbMigrator` project from `3.0.7` to `3.0.8`, which includes changes in the `Version`, `AssemblyVersion`, `FileVersion`, and `PackageVersion` properties in the `EasyDbMigrator.csproj` file. Also, updated the `Polly` package version in the `EasyDbMigrator.csproj` file from `8.3.1` to `8.4.0`. In the `EasyDbMigratorTests.csproj` file, the `Microsoft.NET.Test.Sdk` package version has been updated from `17.9.0` to `17.10.0`. Lastly, the versions of the `xunit` and `xunit.runner.visualstudio` packages in the `EasyDbMigratorTests.csproj` file have been updated from `2.8.0` to `2.8.1`.